requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
firebase_admin
requests
pydantic
python-dotenv